+++ src/react/internal/cache/SuspenseCache.ts
@@ -1,6 +1,6 @@
 import { Trie } from "@wry/trie";
-import type { ObservableQuery } from "../../core/index.js";
-import { canUseWeakMap } from "../../utilities/index.js";
+import type { ObservableQuery } from "../../../core/index.js";
+import { canUseWeakMap } from "../../../utilities/index.js";
 import { InternalQueryReference } from "./QueryReference.js";
 import type { CacheKey } from "./types.js";
