+++ src/react/internal/cache/__tests__/QueryReference.test.ts
